requirements.txt
beautifulsoup4==4.9.3
Flask==2.0.2
googlesearch_python==1.0.1
numpy==1.21.6
pandas==1.3.5
python_dateutil==2.8.2
requests==2.25.1
scikit_learn==1.0.2
whois==0.9.13
gunicorn==20.1.0
Werkzeug==2.0.3

# New libraries you asked to add
selenium==4.1.0
transformers==4.37.2
torch==1.13.1
Pillow==8.4.0
